requirements.txt
# Presentation generation
python-pptx>=0.6.21

# Diagram libraries
diagrams>=0.23.0
graphviz>=0.20.0
networkx>=3.0

# Chart libraries
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Image processing
Pillow>=10.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
python-dateutil>=2.8.0
